requirements.txt
Flask
slack_sdk
python-dotenv
openai
langchain
langchain-openai
gunicorn
requests
colorama